server/pyproject.toml
[tool.poetry]
name = "crm-server"
version = "0.1.0"
description = "Turin Polytechnic University Marketing CRM backend"
authors = ["humoyun-dev <humoyunbektursunniyazov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.29.0"
django-cors-headers = "^4.9.0"
psycopg2-binary = "^2.9.11"
django-filter = "^25.2"
python-dotenv = "^1.2.1"
pytz = "^2025.2"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.8.0"
model-bakery = "^1.18.2"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
